+++ src/Castle.Facilities.NHibernateIntegration.Tests/Internal/CustomConfigurationBuilderTestCase.cs
@@ -1,5 +1,5 @@
 #region License
-// Copyright 2004-2022 Castle Project - https://www.castleproject.org/
+// Copyright 2004-2024 Castle Project - https://www.castleproject.org/
 //
 // Licensed under the Apache License, Version 2.0 (the "License");
 // you may not use this file except in compliance with the License.
@@ -38,6 +38,7 @@ namespace Castle.Facilities.NHibernateIntegration.Tests.Internals
             ConfigurationsCreated++;
 
             var configuration = new DefaultConfigurationBuilder().GetConfiguration(facilityConfiguration);
+
             configuration.Properties["dialect"] =
                 ConfigurationManager.AppSettings["nhf.dialect"];
             configuration.Properties["connection.driver_class"] =