+++ src/Castle.Facilities.NHibernateIntegration/Components/Dao/DataException.cs
@@ -1,5 +1,5 @@
 #region License
-// Copyright 2004-2022 Castle Project - https://www.castleproject.org/
+// Copyright 2004-2024 Castle Project - https://www.castleproject.org/
 //
 // Licensed under the Apache License, Version 2.0 (the "License");
 // you may not use this file except in compliance with the License.
@@ -45,6 +45,7 @@ namespace Castle.Facilities.NHibernateIntegration.Components.Dao
         {
         }
 
+#if NETFRAMEWORK
         /// <summary>
         /// Initializes a new instance of the <see cref="DataException" /> class.
         /// </summary>
@@ -53,5 +54,6 @@ namespace Castle.Facilities.NHibernateIntegration.Components.Dao
         public DataException(SerializationInfo info, StreamingContext context) : base(info, context)
         {
         }
+#endif
     }
 }