+++ CHANGELOG.md
@@ -2,10 +2,10 @@
 
 ## 5.4.0 (2023-xx-xx)
 
-### All
-
 Improvements:
 - Added .NET 7.0 support
+- Updated [System.Data.SqlClient] version to 4.8.5
+- Updated [NHibernate] version to 5.4.0
 
 Breaking Changes:
 - Upgraded [Castle.Core] version to 5.1.0 (TBD)